cleanup and optimizations

+++ base/plugins/iscore-plugin-scenario/Scenario/Document/ScenarioDocument/ScenarioDocumentViewConstants.hpp
@@ -3,8 +3,8 @@
 namespace Scenario
 {
 // SPace at the left of the main box in the main scenario view.
-static const constexpr int ScenarioLeftSpace = 0; // -5
-static const constexpr int ConstraintHeaderHeight = 30;
+static const constexpr double ScenarioLeftSpace = 0.; // -5
+static const constexpr double ConstraintHeaderHeight = 30.;
 
 class ItemType
 {
